Replace data.py with conftest.py
